Update decryption tests
